requirements.txt
django
djangorestframework
requests
beautifulsoup4
gunicorn
django-ninja
django-cors-headers